конфигурирование животных и все классы связанные с этим перенесены в модуль animals-construct-starter starter подключён к основному проекту
