+++ animals-construct-starter/src/main/java/ru/mts/animalsconstructstarter/SetRandomAnimalType.java
@@ -1,0 +1,11 @@
+package ru.mts.animalsconstructstarter;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target({ElementType.METHOD})
+public @interface SetRandomAnimalType {
+}
