+++ animals-construct-starter/src/main/java/ru/mts/animalsconstructstarter/model/FactoryType.java
@@ -1,0 +1,6 @@
+package ru.mts.animalsconstructstarter.model;
+
+public enum FactoryType {
+    PREDATOR_FACTORY,
+    PET_FACTORY
+}
