+++ animals-construct-starter/src/main/java/ru/mts/animalsconstructstarter/model/pet/Parrot.java
@@ -1,0 +1,14 @@
+package ru.mts.animalsconstructstarter.model.pet;
+
+import java.math.BigDecimal;
+import java.time.LocalDate;
+
+public class Parrot extends Pet {
+
+    public final static String PARROT_VOICE = "Кар-кар";
+    public final static String PARROT_BREED = "Попугай";
+
+    public Parrot(String name, BigDecimal cost, String character, LocalDate birthdate) {
+        super(name, cost, character, birthdate, PARROT_BREED, PARROT_VOICE);
+    }
+}
