+++ animals-construct-starter/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,2 @@
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
+  ru.mts.animalsconstructstarter.config.AnimalFactoryServiceConfig, ru.mts.animalsconstructstarter.config.CreateAnimalServiceConfig, ru.mts.animalsconstructstarter.config.AnimalFactoriesConfig
