+++ build.gradle
@@ -17,8 +17,9 @@ repositories {
 }
 
 dependencies {
-    testImplementation 'org.springframework.boot:spring-boot-starter-test'
     implementation 'org.springframework.boot:spring-boot-starter'
+
+    implementation project(':animals-construct-starter')
 }
 
 test {
