+++ settings.gradle
@@ -1,2 +1,2 @@
-
+include ':animals-construct-starter'
