--- src/main/java/ru/mts/SetRandomAnimalType.java
@@ -1,11 +1,0 @@
-package ru.mts;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target({ElementType.METHOD})
-public @interface SetRandomAnimalType {
-}
