--- src/main/java/ru/mts/config/CreateAnimalServiceConfig.java
@@ -1,18 +1,0 @@
-package ru.mts.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Scope;
-import ru.mts.service.AnimalFactoryService;
-import ru.mts.service.CreateAnimalService;
-import ru.mts.service.CreateAnimalServiceImpl;
-
-@Configuration
-public class CreateAnimalServiceConfig {
-
-    @Bean
-    @Scope("prototype")
-    public CreateAnimalService createAnimalService(AnimalFactoryService animalFactoryService) {
-        return new CreateAnimalServiceImpl(animalFactoryService);
-    }
-}
