--- src/main/java/ru/mts/model/FactoryType.java
@@ -1,6 +1,0 @@
-package ru.mts.model;
-
-public enum FactoryType {
-    PREDATOR_FACTORY,
-    PET_FACTORY
-}
