--- src/main/java/ru/mts/service/AnimalFactoryService.java
@@ -1,15 +1,0 @@
-package ru.mts.service;
-
-import ru.mts.factory.AnimalFactory;
-import ru.mts.model.FactoryType;
-
-public interface AnimalFactoryService {
-
-    /**
-     * ищет фабрику указанного типа животного
-     *
-     * @param factoryType тип фабрики
-     * @return фабрику животных
-     */
-    AnimalFactory getFactoryBy(FactoryType factoryType);
-}
