--- src/main/resources/application.yaml
@@ -1,9 +1,0 @@
-logging:
-  level:
-    root: info
-animals:
-  arrays:
-    predatorNames: Мурка, Барсик, Васька, Матроскин, Ляля, Шарик, Рекс, Бобик, Тузик, Дружок, Кеша, Гоша, Чиж, Пушок, Лорд
-    predatorCharacters: Ласковый, игривый, Спокойный, Активный, Умный, Самостоятельный, Жизнерадостный, Терпеливый, Любознательный, Смешной, Храбрый, Сдержанный, Лидер, Послушный, Бесшумный, Спонтанный, Нежный, Заботливый, Агрессивный, Решительный, Беспокойный, Бдительный, Непредсказуемый, Ловкий, Общительный, Независимый, Благородный, Веселый, Серьезный
-    petNames: Мурка, Барсик, Васька, Матроскин, Ляля, Шарик, Рекс, Бобик, Тузик, Дружок, Кеша, Гоша, Чиж, Пушок, Лорд
-    petCharacters: Ласковый, Игривый, Спокойный, Активный, Умный, Самостоятельный, Жизнерадостный, Терпеливый, Любознательный, Смешной, Храбрый, Сдержанный, Лидер, Послушный, Бесшумный, Спонтанный, Нежный, Заботливый, Агрессивный, Решительный, Беспокойный, Бдительный, Непредсказуемый, Ловкий, Общительный, Независимый, Благородный, Веселый, Серьезный
